selenium-tests/requirements.txt
selenium==4.10.0
pytest==7.3.1
pytest-html==3.2.0
pytest-xdist==3.3.1
webdriver-manager==4.0.2
allure-pytest==2.13.2
requests==2.31.0
python-dotenv==1.0.0
faker==18.10.1
assertpy==1.1